chore: refactored exif parser file in order to contain only exif parser code
